Add Integration for Poster and Reporter SDKs
This patch hooks up CI for the Poster and Reporter SDK projects (`/poster` and `/sdk/javascript`). Both take simple Jest tests and have purposefully failing tests right now. That said, the tests should start to ensure we can test could as we build out the projects.

Note: the last piece here before we're basically done the shell is the ability to push poster and/or the SDK to npm. We opt not to do this yet since the projects are still shells.