Manually Set Saddle Build Dir
Saddle expects to find a `solc` build file after compilation. The way we do things, that ends up being in a different directory so that it can come directly from another container's docker build (usually it's `.build/contract.json`). This patch just sets the network config in Saddle to the correct directory, which should solve the current hurdle.