Open Price Feed Poster Clean-up
This patch is simply cleaning up some of the Open Price Feed code. I started and I couldn't stop, so it's a nice overhaul just trying to cleanup the functions and interface without significantly changing the behavior. There are some minor CLI differences that we'll need to account for. This is going to be the basis for the branch to handle updating Uniswap for testnets. We do change the CLI and some minor features, but I think this is coupled with some pretty serious bugfixes, as well.