Add Compound Config
This patch adds Compound Config to the Open Oracle that allows us to trac the Compound Config repo. This means, for instance, when you run a `saddle console -n mainnet`, you'll actually have access to the Open Price Feed contracts. This makes it easy to get a console and interact with the OPF.